chore: separate util type adapters from the main type adapter factory and remove the main type adapter factory
