+++ data/json/src/main/java/net/hypejet/jet/data/json/adapters/util/UtilTypeAdapterFactory.java
@@ -1,33 +1,31 @@
-package net.hypejet.jet.data.json;
+package net.hypejet.jet.data.json.adapters.util;
 
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
 import com.google.gson.TypeAdapterFactory;
 import com.google.gson.reflect.TypeToken;
-import net.hypejet.jet.data.json.adapters.UnitTypeAdapter;
 import net.hypejet.jet.data.json.util.JsonUnit;
 
 /**
  * Represents a {@linkplain TypeAdapterFactory type adapter factory} providing {@linkplain TypeAdapter type adapters}
- * converting Jet data objects.
+ * of utility objects.
  *
  * @since 1.0
  * @see TypeAdapterFactory
  */
-public final class DataTypeAdapterFactory implements TypeAdapterFactory {
+public final class UtilTypeAdapterFactory implements TypeAdapterFactory {
     /**
-     * An instance of the {@linkplain DataTypeAdapterFactory data type adapter factory}.
+     * An instance of the {@linkplain UtilTypeAdapterFactory util type adapter factory}.
      *
      * @since 1.0
      */
-    public static final DataTypeAdapterFactory INSTANCE = new DataTypeAdapterFactory();
+    public static final UtilTypeAdapterFactory INSTANCE = new UtilTypeAdapterFactory();
 
-    private DataTypeAdapterFactory() {}
+    private UtilTypeAdapterFactory() {}
 
     @Override
     public <T> TypeAdapter create(Gson gson, TypeToken<T> type) {
-        Class<? super T> rawType = type.getRawType();
-        if (JsonUnit.class.isAssignableFrom(rawType)) {
+        if (JsonUnit.class.isAssignableFrom(type.getRawType())) {
             return UnitTypeAdapter.INSTANCE;
         } else {
             return null;
